Update for propose of file change
Joining as a Member or other participant means joining a global network of individuals and organizations that work together, in many different ways, to promote and enable openness and sharing around the world. We are all collectively committed to building a global network that is supportive and that recognises and celebrates the achievements of others. I expect that as a Member or other participant, you will strive to work collaboratively and help others achieve our shared goals.

In my activities as a Member, many Members will work with others in your country to constitute a local Chapter, organizing activities consistent and aligned with the CCGN shared strategic vision. In addition to a local Chapter’s organizational structure as described above, CCGN Platforms frame what a Member or other contributors can say and do on behalf of Creative Commons or the local Chapter. Platforms enable diverse